tools/requirements.txt
paho-mqtt
pandas
psycopg2-binary
pyarrow
fastparquet